+++ src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -2,8 +2,8 @@ package ru.practicum.shareit.booking;
 
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.item.Item;
+import ru.practicum.shareit.user.User;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
