+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -2,8 +2,6 @@ package ru.practicum.shareit.booking;
 
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.booking.dto.BookingDto;
-import ru.practicum.shareit.booking.dto.BookingDtoToCreate;
 
 import java.util.List;
