+++ src/main/java/ru/practicum/shareit/booking/BookingDto.java
@@ -1,0 +1,30 @@
+package ru.practicum.shareit.booking;
+
+import lombok.Builder;
+import lombok.Data;
+import ru.practicum.shareit.item.Item;
+import ru.practicum.shareit.user.User;
+
+import javax.validation.constraints.FutureOrPresent;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+
+@Data
+@Builder
+public class BookingDto {
+
+    private long id;
+
+    @NotNull
+    private LocalDateTime start;
+    @NotNull
+    @FutureOrPresent
+    private LocalDateTime end;
+
+    private Item item;
+
+    private User booker;
+
+    private BookingStatus status;
+}
