+++ src/main/java/ru/practicum/shareit/booking/BookingDtoForItemHost.java
@@ -1,0 +1,19 @@
+package ru.practicum.shareit.booking;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class BookingDtoForItemHost {
+
+    private long id;
+
+    private LocalDateTime start;
+
+    private LocalDateTime end;
+
+    private long bookerId;
+}
