+++ src/main/java/ru/practicum/shareit/booking/BookingDtoToCreate.java
@@ -1,0 +1,23 @@
+package ru.practicum.shareit.booking;
+
+import lombok.Builder;
+import lombok.Data;
+
+import javax.validation.constraints.FutureOrPresent;
+import javax.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class BookingDtoToCreate {
+    @NotNull
+    @FutureOrPresent
+    private LocalDateTime start;
+
+    @NotNull
+    @FutureOrPresent
+
+    private LocalDateTime end;
+
+    private long itemId;
+}
