+++ src/main/java/ru/practicum/shareit/booking/BookingMapper.java
@@ -1,0 +1,49 @@
+package ru.practicum.shareit.booking;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class BookingMapper {
+
+
+    public static BookingDto modelToDto(Booking booking) {
+        return BookingDto.builder()
+                .id(booking.getId())
+                .start(booking.getStart())
+                .end(booking.getEnd())
+                .item(booking.getItem())
+                .booker(booking.getBooker())
+                .status(booking.getStatus())
+                .build();
+    }
+
+    public static Booking dtoToCreateToModel(BookingDtoToCreate bookingDto) {
+        return Booking.builder()
+                .end(bookingDto.getEnd())
+                .start(bookingDto.getStart())
+                .build();
+    }
+
+
+    public static Booking dtoToModel(BookingDto bookingDto) {
+        return Booking.builder()
+                .id(bookingDto.getId())
+                .start(bookingDto.getStart())
+                .end(bookingDto.getEnd())
+                .item(bookingDto.getItem())
+                .booker(bookingDto.getBooker())
+                .status(bookingDto.getStatus())
+                .build();
+    }
+
+    public static BookingDtoForItemHost modelToDtoForItem(Booking booking) {
+        return BookingDtoForItemHost.builder()
+                .id(booking.getId())
+                .bookerId(booking.getId())
+                .start(booking.getStart())
+                .end(booking.getEnd())
+                .bookerId(booking.getBooker().getId())
+                .build();
+    }
+}
