+++ src/main/java/ru/practicum/shareit/booking/BookingRepository.java
@@ -16,7 +16,7 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
     @Modifying
     void updateStatus(BookingStatus status, long id);
 
-    public List<Booking> findAllByBookerId(long bookerId, Sort sort);
+    List<Booking> findAllByBookerId(long bookerId, Sort sort);
 
     @Query("select b from Booking as b " +
             "join b.booker as u  " +
@@ -84,10 +84,25 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
             "and b.status = 'WAITING' ")
     List<Booking> findAllItemsWaitingBookings(List userItems, Sort sort);
 
+    /*@Query("select b from Booking as b " +
+            "join b.item as i " +
+            "where i.id = ?1 and " +
+            "b.status = 'APPROVED' and " +
+            "(b.end < ?2 or b.start > ?2)")
+    List<Booking> findAllBookingsByItemId(long itemId, LocalDateTime dateTime);*/
+
     @Query("select b from Booking as b " +
             "join b.item as i " +
-            "where i.id = ?1 and" +
-            "(b.end < ?2 or b.start > ?2)")
-    List<Booking> findAllBookingsByItemId(long itemId, LocalDateTime dateTime);
+            "where i.id = ?1 and " +
+            "b.status = 'APPROVED'")
+    List<Booking> findAllBookingsByItemId(long itemId);
+
+    @Query("select b from Booking as b " +
+            "join b.item as i " +
+            "join b.booker as u " +
+            "where i.id = ?1 " +
+            "and u.id = ?2 " +
+            "and b.end < ?3 ")
+    List<Booking> findAllByBookerIdAndItemIdAndEndBefore(long itemId, long bookerId,  LocalDateTime end);
 
 }
