+++ src/main/java/ru/practicum/shareit/booking/BookingService.java
@@ -1,8 +1,5 @@
 package ru.practicum.shareit.booking;
 
-import ru.practicum.shareit.booking.dto.BookingDto;
-import ru.practicum.shareit.booking.dto.BookingDtoToCreate;
-
 import java.util.List;
 
 public interface BookingService {
