--- src/main/java/ru/practicum/shareit/booking/dto/BookingDtoForItemHost.java
@@ -1,19 +1,0 @@
-package ru.practicum.shareit.booking.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class BookingDtoForItemHost {
-
-    private long id;
-
-    private LocalDateTime start;
-
-    private LocalDateTime end;
-
-    private long bookerId;
-}
