+++ src/main/java/ru/practicum/shareit/item/CommentDtoToCreate.java
@@ -1,0 +1,22 @@
+package ru.practicum.shareit.item;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CommentDtoToCreate {
+
+    @NotBlank
+    @Size(max = 1000)
+    private String text;
+
+
+}
