+++ src/main/java/ru/practicum/shareit/item/CommentDtoToReturn.java
@@ -1,0 +1,20 @@
+package ru.practicum.shareit.item;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class CommentDtoToReturn {
+
+    private long id;
+
+    private String text;
+
+    private String authorName;
+
+    private LocalDateTime created;
+
+}
