+++ src/main/java/ru/practicum/shareit/item/ItemRepository.java
@@ -1,0 +1,20 @@
+package ru.practicum.shareit.item;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+
+import java.util.List;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+
+    List<Item> findAllByOwnerIdOrderByIdAsc(long userId);
+
+
+    @Query("select it " +
+            "from Item as it " +
+            "where ((lower(it.name) like %:text% " +
+            "or lower(it.description) like %:text%)" +
+            "and it.isAvailable is true) ")
+    List<Item> findAllByText(@Param("text") String text);
+}
