--- src/main/java/ru/practicum/shareit/item/dto/mapper/ItemMapper.java
@@ -1,39 +1,0 @@
-package ru.practicum.shareit.item.dto.mapper;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import ru.practicum.shareit.item.dto.ItemDto;
-import ru.practicum.shareit.item.dto.ItemDtoWithBookings;
-import ru.practicum.shareit.item.model.Item;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class ItemMapper {
-
-    public static ItemDto modelToDto(Item item) {
-        return ItemDto.builder()
-                .id(item.getId())
-                .name(item.getName())
-                .description(item.getDescription())
-                .available(item.getIsAvailable())
-                .build();
-    }
-
-    public static Item dtoToModel(ItemDto itemDto) {
-        return Item.builder()
-                .id(itemDto.getId())
-                .name(itemDto.getName())
-                .description(itemDto.getDescription())
-                .isAvailable(itemDto.getAvailable())
-                .build();
-    }
-
-    public static ItemDtoWithBookings modelToDtoWithBookings(Item item) {
-        return ItemDtoWithBookings.builder()
-                .id(item.getId())
-                .name(item.getName())
-                .description(item.getDescription())
-                .available(item.getIsAvailable())
-                .build();
-
-    }
-}
