--- src/main/java/ru/practicum/shareit/request/dto/ItemRequestDto.java
@@ -1,29 +1,0 @@
-package ru.practicum.shareit.request.dto;
-
-import lombok.Builder;
-import lombok.Data;
-import ru.practicum.shareit.user.model.User;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-import java.time.LocalDateTime;
-
-/**
- * TODO Sprint add-item-requests.
- */
-
-@Data
-@Builder
-public class ItemRequestDto {
-
-    @NotBlank
-    @Size(max = 500)
-    private String description;
-
-    @NotNull
-    private User requestor;
-
-    private LocalDateTime created;
-
-}
