+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,0 +1,18 @@
+package ru.practicum.shareit.user;
+
+import java.util.List;
+
+public interface UserService {
+
+    UserDto getUserById(long userId);
+
+    List<UserDto> getAllUsers();
+
+    UserDto createNewUser(UserDto userDto);
+
+    UserDto updateUser(long userId, UserDto userDto);
+
+    void deleteUserById(long userId);
+
+    void deleteUsers();
+}
