+++ src/main/java/ru/practicum/shareit/user/UserShort.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.user;
+
+public interface UserShort {
+    String getId();
+    String getName();
+    String getEmail();
+}
