--- src/main/java/ru/practicum/shareit/user/dto/mapper/UserMapper.java
@@ -1,26 +1,0 @@
-package ru.practicum.shareit.user.dto.mapper;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import ru.practicum.shareit.user.dto.UserDto;
-import ru.practicum.shareit.user.model.User;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UserMapper {
-
-    public static UserDto modelToDto(User user) {
-        return UserDto.builder()
-                .id(user.getId())
-                .name(user.getName())
-                .email(user.getEmail())
-                .build();
-    }
-
-    public static User dtoToModel(UserDto userDto) {
-        return User.builder()
-                .id(userDto.getId())
-                .name(userDto.getName())
-                .email(userDto.getEmail())
-                .build();
-    }
-}
