--- src/main/java/ru/practicum/shareit/user/model/UserShort.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.user.model;
-
-public interface UserShort {
-    String getId();
-    String getName();
-    String getEmail();
-}
